fix test.sh locally and hopefully in actions

+++ test.sh
@@ -4,12 +4,13 @@ home-manager switch
 # echo "running emacs: $(which emacs)"
 # Before running tests, clean the stack directory for our haskell test project
 pushd testdata/simple-haskell-project && stack clean && popd
-EMACSFOR="PERSONAL" emacs -nw --load load-init-then-run-ert.el
+EMACSFOR="PERSONAL" emacs -batch -f package-initialize -L . -f buttercup-run-discover
 emacsExitCode=$?;
-cat test-results.txt
-echo "github env vars: "
-env | grep -i git
-exit $emacsExitCode;
+
+if [ -z "$GITHUB_ACTIONS" ]; then
+    exit $emacsExitCode;
+fi
 if [ -f "test-results.txt" ]; then
+    cat test-results.txt
     rm test-results.txt
 fi
